Fetch user from server, set locale in realm and change language with i18next
