+++ src/realmModels/index.js
@@ -18,7 +18,7 @@ export default {
     Taxon,
     User
   ],
-  schemaVersion: 24,
+  schemaVersion: 25,
   path: "db.realm",
   migration: ( oldRealm, newRealm ) => {
     if ( oldRealm.schemaVersion < 21 ) {
