+++ tests/factories/LocalUser.js
@@ -3,5 +3,6 @@ import { define } from "factoria";
 export default define( "LocalUser", faker => ( {
   login: faker.internet.userName( ),
   id: faker.datatype.number( ),
-  iconUrl: faker.image.imageUrl( )
+  iconUrl: faker.image.imageUrl( ),
+  locale: faker.random.locale( )
 } ) );
